automation/manifest-based-boardfarm-run.groovy: match job name and build number
Reviewed-on: https://cicd.lgirdk.com/gerrit/c/github/boardfarm/+/3030
Tested-by: Jenkins Automation <jenkins@libertyglobal.com>

+++ automation/manifest-based-boardfarm-run.groovy
@@ -323,11 +323,14 @@ pipeline {
             for PID in $(ps -eo pid,command -u ${USER} | grep -v grep | tail -n+2 | awk '{print $1}' | grep -v ${PID_SELF} | grep -v ${PPID}); do
                 echo "Checking pid ${PID}"
                 if xargs -0 -L1 -a /proc/${PID}/environ 2>/dev/null | grep "BUILD_ID=${BUILD_ID}$"; then
-                    echo "Matched BUILD_ID=${BUILD_ID}$"
-                    echo "Killing $(ps -p ${PID} | tail -1 | awk '{print $1}')"
-                    sed -z 's/$/ /' /proc/$PID/cmdline; echo
-                    kill ${PID}
-                    echo killed ${PID}
+                    if xargs -0 -L1 -a /proc/${PID}/environ 2>/dev/null | grep "JOB_NAME=${JOB_NAME}$"; then
+                        echo "Matched BUILD_ID=${BUILD_ID}$"
+                        echo "Matched JOB_NAME=${JOB_NAME}$"
+                        echo "Killing $(ps -p ${PID} | tail -1 | awk '{print $1}')"
+                        sed -z 's/$/ /' /proc/$PID/cmdline; echo
+                        kill ${PID}
+                        echo killed ${PID}
+                    fi
                 fi
             done || true
             '''
